requirements.txt
streamlit 
pandas 
numpy
matplotlib.pyplot
plotly.graph_objects
os
